+++ core/src/test/java/com/graphhopper/routing/util/FootFlagEncoderTest.java
@@ -250,7 +250,7 @@ public class FootFlagEncoderTest {
         way.setTag("route", "ferry");
         // a bit longer than an hour
         way.setTag("duration:seconds", 4000L);
-        way.setTag("road_distance", 30000.0);
+        way.setTag("edge_distance", 30000.0);
         // the speed is truncated to maxspeed (=15)
         assertEquals(15, footEncoder.ferrySpeedCalc.getSpeed(way), .1);
         IntsRef flags = footEncoder.handleWayTags(encodingManager.createEdgeFlags(), way);